+++ src/controllers/media.ts
@@ -126,38 +126,46 @@ export const getSeriesV2 = async(ctx: ParameterizedContext): Promise<Partial<Ser
  * we use that has that functionality.
  */
 export const getSeason = async(ctx: ParameterizedContext): Promise<Partial<SeasonMetadataInterface>> => {
-  try {
-    const { tmdbID, season, title, language, year }: UmsQueryParams = ctx.query;
-    if (!tmdbID && !title) {
-      throw new ValidationError('title or tmdbID is required');
-    }
-    if (!season) {
-      throw new ValidationError('season is required');
-    }
-    const seasonNumber = Number(season);
-    let tmdbTvID = tmdbID;
+  const { season, title, year }: UmsQueryParams = ctx.query;
+  if (!title || !season) {
+    throw new ValidationError('title and season are required');
+  }
+  const seasonNumber = Number(season);
 
-    if (!tmdbTvID) {
-      //todo: add language when available
-      const seriesMetadata = await externalAPIHelper.getSeriesMetadata(null, title, year);
-      if (!seriesMetadata?.tmdbID) {
-        throw new MediaNotFoundError();
-      }
-      tmdbTvID = seriesMetadata.tmdbID;
-    }
-
-    const seasonMetadata = await externalAPIHelper.getSeasonMetadata(tmdbTvID, seasonNumber);
-    if (_.isEmpty(seasonMetadata)) {
-      throw new MediaNotFoundError();
-    }
-    return ctx.body = seasonMetadata;
-  } catch (err) {
-    // log unexpected errors
-    if (!(err instanceof MediaNotFoundError)) {
-      console.error(err);
-    }
+  // Return early for previously-failed lookups
+  const failedLookupQuery: FailedLookupsInterface = { title, season, type: 'season' };
+  if (year) {
+    failedLookupQuery.year = year;
+  }
+  if (await FailedLookups.findOne(failedLookupQuery, '_id', { lean: true }).exec()) {
+    await FailedLookups.updateOne(failedLookupQuery, { $inc: { count: 1 } }).exec();
     throw new MediaNotFoundError();
   }
+
+  const seriesMetadata = await externalAPIHelper.getSeriesMetadata(null, title, year);
+  if (!seriesMetadata?.tmdbID) {
+    await FailedLookups.updateOne(failedLookupQuery, { $inc: { count: 1 } }, { upsert: true, setDefaultsOnInsert: true }).exec();
+    throw new MediaNotFoundError();
+  }
+
+  // Start TMDB lookups
+  const seasonRequest = {
+    append_to_response: 'images,external_ids,credits',
+    id: seriesMetadata.tmdbID,
+    season_number: seasonNumber,
+  };
+
+  const tmdbResponse = await tmdb.seasonInfo(seasonRequest);
+  const tmdbData = mapper.parseTMDBAPISeasonResponse(tmdbResponse);
+  // End TMDB lookups
+
+  if (_.isEmpty(tmdbData)) {
+    await FailedLookups.updateOne(failedLookupQuery, { $inc: { count: 1 } }, { upsert: true, setDefaultsOnInsert: true }).exec();
+    throw new MediaNotFoundError();
+  }
+
+  const seasonMetadata = await SeasonMetadata.create(tmdbData);
+  return ctx.body = seasonMetadata;
 };
 
 export const getVideoV2 = async(ctx: ParameterizedContext): Promise<MediaMetadataInterface> => {
