+++ src/models/SeasonMetadata.ts
@@ -12,7 +12,6 @@ export interface SeasonMetadataInterface {
   posterRelativePath?: string;
   seasonNumber: number;
   tmdbID?: number;
-  tmdbTvID?: number;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -26,9 +25,8 @@ const SeasonMetadataSchema: Schema = new Schema({
   name: { type: String },
   overview: { type: String },
   posterRelativePath: { type: String },
-  seasonNumber: { type: Number, index: true, required: true },
+  seasonNumber: { type: Number },
   tmdbID: { type: Number },
-  tmdbTvID: { type: Number, index: true, required: true },
 }, {
   collection: 'season_metadata',
   timestamps: true,
